refactor: move ir decoding helpers to separate header / implementation file.
