+++ .gitignore
@@ -21,3 +21,4 @@ settings.json
 CMakeSettings.json
 CppProperties.json
 result
+.cache/
