+++ src/library/compiler/CMakeLists.txt
@@ -5,5 +5,5 @@ add_library(compiler OBJECT init_module.cpp ## New
   reduce_arity.cpp closed_term_cache.cpp
   export_attribute.cpp extern_attribute.cpp
   borrowed_annotation.cpp init_attribute.cpp eager_lambda_lifting.cpp
-  struct_cases_on.cpp find_jp.cpp ir.cpp implemented_by_attribute.cpp
+  struct_cases_on.cpp find_jp.cpp ir.cpp ir_decode_helpers.cpp implemented_by_attribute.cpp
   ir_interpreter.cpp)
